pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "help-exilium"
version = "0.1.0"
description = "Bot Discord do Exilium"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py==2.3.2",
    "python-dotenv==1.0.1",
]

